requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.3.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.2.0
langsmith>=0.1.0
langgraph-cli>=0.0.40
pydantic>=2.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
exa-py>=1.0.0
llama-parse>=0.1.0
pinecone-client>=3.0.0
e2b-code-interpreter>=0.1.0
matplotlib>=3.8.0
pandas>=2.0.0
typing-extensions>=4.8.0
markdown>=3.4.0


weasyprint>=60.0.0
